node: move default apps into config & network into constants
